Rewrote all of the select fields to extend a modified version of the standard select field to make code reusable and and to add consistency to the rendering of all of the different select/multi-select fields.
Also added an attribute "multiselect" to the base class, which allows a drop-down or multi-select menu to be chosen for ANY of the select fields by simply adding a new attribute in options.php. This reduces the number of additional field classes that need to be added for every different type of select list. So instead of tags select and tags select multi, in theory there is only the need for a single tags select class and then the user sets the multiselect attribute to true or false depending on whether a dropdown or multi-select is desired.

Granted at this point this is rather moot since there are dropdown and multi-select fields for most everything...but it's still good practice in case the base class needs to be modified for any reason (e.g. styling).

NOTE: I didn't include the extra <br /> tag between the list and the description for multi-select lists, but this is a trivial modification on the base select class that I might add later.